feat/process recommendations

+++ app/src/main/java/com/akilimo/mobile/enums/EnumUnitOfSale.kt
@@ -8,14 +8,14 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 enum class EnumUnitOfSale(
-    val weight: Int,
+    val weight: Double,
     val isUniversal: Boolean,
     private val labelRes: Int,
     private val textRes: Int
 ) : Parcelable, ILabelProvider {
     // Crop-specific unit (only maize fresh cob)
     FRESH_COB(
-        weight = 1,
+        weight = 1.0,
         isUniversal = false,
         labelRes = R.string.lbl_fresh_cob,
         textRes = R.string.lbl_fresh_cob
@@ -23,37 +23,37 @@ enum class EnumUnitOfSale(
 
     // Universal units (apply to all produce types)
     ONE_KG(
-        weight = 1,
+        weight = 1.0,
         isUniversal = true,
         labelRes = R.string.lbl_one_kg_unit,
         textRes = R.string.per_kg_sale_unit
     ),
     FIFTY_KG(
-        weight = 50,
+        weight = 50.0,
         isUniversal = true,
         labelRes = R.string.lbl_50_kg_unit,
         textRes = R.string.per_50_kg_sale_unit
     ),
     HUNDRED_KG(
-        weight = 100,
+        weight = 100.0,
         isUniversal = true,
         labelRes = R.string.lbl_100_kg_unit,
         textRes = R.string.per_100_kg_sale_unit
     ),
     THOUSAND_KG(
-        weight = 1000,
+        weight = 1000.0,
         isUniversal = true,
         labelRes = R.string.lbl_1000_kg_unit,
         textRes = R.string.per_tonne_sale_unit
     ),
     PICKUP_LOAD(
-        weight = 1,
+        weight = 1.0,
         isUniversal = false,
         labelRes = R.string.per_pick_up_load_sale_unit,
         textRes = R.string.per_pick_up_load_sale_unit
     );
 
-    fun unitWeight(): Int = weight
+    fun unitWeight(): Double = weight
 
     fun unitOfSale(context: Context): String =
         labelRes.let(context::getString)
